pyproject.toml
[project]
name = "advertising-dashboard"
version = "0.1.0"
description = "Advertising Dashboard MVP with Streamlit and OpenRouter.ai integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.4",
    "streamlit>=1.45.0",
    "python-dotenv>=1.0.1",
    "passlib>=1.7.4",
    "httpx>=0.27.0",
    "pillow>=10.3.0",
    "pandas>=2.2.1",
    "altair>=5.3.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)s] %(filename)s:%(lineno)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
testpaths = ["tests"]

[tool.pytest-env]
MODE = "dev"

[tool.pytest-asyncio]
asyncio_mode = "auto"
default_test_loop_scope = "session"
default_fixture_loop_scope = "session"

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "D",
  "S311",
  "ANN401",
  "PGH003",
  "PTH123",
  "PTH118",
  "TRY003",
  "EM101",
  "EM102",
  "FBT003",
  "FBT002",
  "FBT001",
  "TC001",
  "TC002",
  "TD002",
  "TD003",
  "FIX002",
  "G004",
  "B008"
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.per-file-ignores]
"**/{tests,docs,tools}/*" = ["S101", "SLF001", "ANN", "ARG", "PTH", "PERF", "PLR2004"]
